Apply the separation of concerns
